first working scf workchain tests, still minor improvements needed
