+++ aiida_fleur/tests/workflows/calc_data_dir/mock-fleur-15e765f12fcdbe12c715120ecf8625ad/out.error
@@ -1,0 +1,12 @@
+I/O warning : failed to load external entity "relax.xml"
+ 
+ *****************************************
+ Run finished successfully
+ Stop message:
+   all done
+ *****************************************
+Rank:0 used    0.699GB/	  767896 kB
+  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
+                                 Dload  Upload   Total   Spent    Left  Speed
+  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0100   782  100    40  100   742    166   3085 --:--:-- --:--:-- --:--:--  3091
+OK
