+++ aiida_fleur/tests/workflows/calc_data_dir/mock-fleur-58394cd5422cd3fc7a6d802d966c2848/out.error
@@ -1,0 +1,12 @@
+I/O warning : failed to load external entity "relax.xml"
+ 
+ *****************************************
+ Run finished successfully
+ Stop message:
+   all done
+ *****************************************
+Rank:0 used    0.672GB/	  704468 kB
+  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
+                                 Dload  Upload   Total   Spent    Left  Speed
+  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0100   781  100    40  100   741    211   3915 --:--:-- --:--:-- --:--:--  3900
+OK
