+++ aiida_fleur/tests/workflows/calc_data_dir/mock-fleur-58394cd5422cd3fc7a6d802d966c2848/shell.out
@@ -1,0 +1,8 @@
+      Welcome to FLEUR        (www.flapw.de)   
+      MaX-Release 4.0          (www.max-centre.eu)
+  stars are always ordered 
+ --------------------------------------------------------
+ Number of OMP-threads:           6
+ --------------------------------------------------------
+ Iteration:           1  Distance:  6.340901440083457E-006
+ Usage data send using curl: usage.json
