+++ aiida_fleur/tests/workflows/calc_data_dir/mock-inpgen-4476cada582e637fef7432a68c19d42a/aiida.in
@@ -1,0 +1,17 @@
+A Fleur input generator calculation with aiida
+&input  cartesian=F /
+      0.0000000000       5.1673552979       5.1673552979
+      5.1673552979       0.0000000000       5.1673552979
+      5.1673552979       5.1673552979       0.0000000000
+      1.0000000000
+      1.0000000000       1.0000000000       1.0000000000
+
+      2
+         14       0.1250000000       0.1250000000       0.1250000000
+         14      -0.1250000000      -0.1250000000      -0.1250000000
+&atom
+  element="Si"   jri=981   lmax=12   lnonsph=6   rmt=2.1 /
+&comp
+  kmax=3.2 /
+&kpt
+  div1=10   div2=10   div3=10   tkb=0.0005 /
