+++ aiida_fleur/tests/workflows/calc_data_dir/mock-inpgen-4476cada582e637fef7432a68c19d42a/shell.out
@@ -1,0 +1,2 @@
+      Welcome to FLEUR - inpgen   (www.flapw.de)   
+      MaX-Release 4.0          (www.max-centre.eu)
