+++ aiida_fleur/tests/workflows/calc_data_dir/mock-inpgen-4476cada582e637fef7432a68c19d42a/struct.xsf
@@ -1,0 +1,10 @@
+ CRYSTAL
+ PRIMVEC
+.0000000 2.7344467 2.7344467
+2.7344467 .0000000 2.7344467
+2.7344467 2.7344467 .0000000
+ PRIMCOORD
+           2  1
+  14  .6836117 .6836117 .6836117
+  14  -.6836117 -.6836117 -.6836117
+
